requirements.txt
python-telegram-bot[rate-limiter]==21.0
openai==1.13.3
tiktoken>=0.5.1
PyYAML==6.0.1
pymongo==4.5.0
python-dotenv==1.0.1
watchdog==2.3.1
prisma==0.12.0